add button: layers & widgets
