+++ lib/assets/core/javascripts/cartodb3/editor/editor-pane.js
@@ -3,9 +3,6 @@ var Backbone = require('backbone');
 var _ = require('underscore');
 var CoreView = require('backbone/core-view');
 var createTextLabelsTabPane = require('../components/tab-pane/create-text-labels-tab-pane');
-var AddWidgetsView = require('../components/modals/add-widgets/add-widgets-view');
-var AddLayerView = require('../components/modals/add-layer/add-layer-view');
-var AddLayerModel = require('../components/modals/add-layer/add-layer-model');
 var StackLayoutView = require('../components/stack-layout/stack-layout-view');
 var Header = require('./editor-header.js');
 var EditorTabPaneTemplate = require('./editor-tab-pane.tpl');
@@ -46,10 +43,6 @@ var MAX_LAYERS_REACHED = 'MAX_LAYERS_REACHED';
 module.exports = CoreView.extend({
 
   className: 'Editor-content',
-
-  events: {
-    'click .js-add': '_addItem'
-  },
 
   initialize: function (opts) {
     checkAndBuildOpts(opts, REQUIRED_OPTS, this);
@@ -100,6 +93,7 @@ module.exports = CoreView.extend({
                       userModel: self._userModel,
                       editorModel: self._editorModel,
                       configModel: self._configModel,
+                      pollingModel: self._pollingModel,
                       userActions: self._userActions,
                       layerDefinitionsCollection: self._layerDefinitionsCollection,
                       analysisDefinitionNodesCollection: self._analysisDefinitionNodesCollection,
@@ -147,6 +141,9 @@ module.exports = CoreView.extend({
                       modals: self._modals,
                       layerDefinitionsCollection: self._layerDefinitionsCollection,
                       widgetDefinitionsCollection: self._widgetDefinitionsCollection,
+                      analysisDefinitionNodesCollection: self._analysisDefinitionNodesCollection,
+                      userModel: self._userModel,
+                      configModel: self._configModel,
                       stackLayoutModel: self._mapStackLayoutModel
                     });
                   }
@@ -438,52 +435,6 @@ module.exports = CoreView.extend({
         // TODO: trigger element creation
         break;
     }
-  },
-
-  _addItem: function () {
-    // We have to manually ignore these events because a disabled button won't trigger mouseover either => no tooltips
-    if (this.$('.js-add').hasClass('is-disabled')) return;
-    switch (this._mapTabPaneView.getSelectedTabPaneName()) {
-      case LAYERS_TAB_NAME: return this._addLayer();
-      case 'widgets': return this._addWidget();
-    }
-  },
-
-  _addLayer: function () {
-    var self = this;
-    var modal = this._modals.create(function (modalModel) {
-      var addLayerModel = new AddLayerModel({}, {
-        userModel: self._userModel,
-        userActions: self._userActions,
-        configModel: self._configModel,
-        pollingModel: self._pollingModel
-      });
-
-      return new AddLayerView({
-        modalModel: modalModel,
-        configModel: self._configModel,
-        userModel: self._userModel,
-        createModel: addLayerModel,
-        pollingModel: self._pollingModel
-      });
-    });
-    modal.show();
-  },
-
-  _addWidget: function () {
-    var self = this;
-
-    this._modals.create(function (modalModel) {
-      return new AddWidgetsView({
-        modalModel: modalModel,
-        userModel: self._userModel,
-        userActions: self._userActions,
-        configModel: self._configModel,
-        analysisDefinitionNodesCollection: self._analysisDefinitionNodesCollection,
-        layerDefinitionsCollection: self._layerDefinitionsCollection,
-        widgetDefinitionsCollection: self._widgetDefinitionsCollection
-      });
-    });
   },
 
   _share: function () {
