+++ lib/assets/core/javascripts/cartodb3/editor/editor-tab-pane.tpl
@@ -1,9 +1,4 @@
 <nav class="CDB-NavMenu js-theme">
   <ul class="CDB-NavMenu-inner CDB-Text is-semibold CDB-Size-medium js-menu"></ul>
-  <button class="Editor-buttonNavigation CDB-Button CDB-Button--small CDB-Button--primary js-add">
-    <span class="CDB-Button-Text CDB-Text is-semibold CDB-Size-small u-upperCase">
-      <%- _t('editor.button_add') %>
-    </span>
-  </button>
 </nav>
 <div class="Editor-content js-content"></div>
