+++ lib/assets/core/javascripts/cartodb3/editor/layers/layers-view.js
@@ -6,6 +6,9 @@ var Notifier = require('../../components/notifier/notifier');
 var template = require('./layers.tpl');
 var LayerViewFactory = require('./layer-view-factory');
 var checkAndBuildOpts = require('../../helpers/required-opts');
+var TipsyTooltipView = require('../../components/tipsy-tooltip-view');
+var AddLayerView = require('../../components/modals/add-layer/add-layer-view');
+var AddLayerModel = require('../../components/modals/add-layer/add-layer-model');
 
 var SORTABLE_SELECTOR = '.js-layers';
 var SORTABLE_ITEMS_SELECTOR = '.js-layer.js-sortable-item';
@@ -17,6 +20,7 @@ var REQUIRED_OPTS = [
   'modals',
   'userModel',
   'configModel',
+  'pollingModel',
   'editorModel',
   'userActions',
   'stateDefinitionModel',
@@ -29,6 +33,10 @@ var REQUIRED_OPTS = [
  * View to render layer definitions list
  */
 module.exports = CoreView.extend({
+
+  events: {
+    'click .js-add': '_addLayer'
+  },
 
   initialize: function (opts) {
     checkAndBuildOpts(opts, REQUIRED_OPTS, this);
@@ -60,8 +68,44 @@ module.exports = CoreView.extend({
 
     _.each(this._layerDefinitionsCollection.toArray().reverse(), this._addLayerView, this);
     this._initSortable();
+    this._initViews();
 
     return this;
+  },
+
+  _initViews: function () {
+    var tooltip = new TipsyTooltipView({
+      el: this.$('.js-add'),
+      gravity: 'w',
+      title: function () {
+        return _t('editor.layers.add-layer.tooltip');
+      },
+      offset: 8
+    });
+    this.addView(tooltip);
+  },
+
+  _addLayer: function () {
+    if (this.$('.js-add').hasClass('is-disabled')) return;
+
+    var self = this;
+    var modal = this._modals.create(function (modalModel) {
+      var addLayerModel = new AddLayerModel({}, {
+        userModel: self._userModel,
+        userActions: self._userActions,
+        configModel: self._configModel,
+        pollingModel: self._pollingModel
+      });
+
+      return new AddLayerView({
+        modalModel: modalModel,
+        configModel: self._configModel,
+        userModel: self._userModel,
+        createModel: addLayerModel,
+        pollingModel: self._pollingModel
+      });
+    });
+    modal.show();
   },
 
   _createNotification: function (m) {
