+++ lib/assets/core/javascripts/cartodb3/editor/layers/layers.tpl
@@ -1,2 +1,5 @@
+<button class="CDB-Button CDB-Button--dashed CDB-Button--wide CDB-Button-Text CDB-Text is-semibold CDB-Size-small u-upperCase u-bSpace--m js-add">
+  + <%- _t('editor.layers.add-layer.label') %>
+</button>
 <ul class="Editor-ListLayer js-layers"></ul>
 <div class="Editor-dropArea js-drop"></div>
