+++ lib/assets/core/javascripts/cartodb3/editor/widgets/widgets-placeholder.tpl
@@ -1,13 +1,4 @@
-<h2 class="CDB-Text CDB-Size-huge is-light u-secondaryTextColor"><%- _t('editor.widgets.widgets-form.placeholder-text') %></h2>
-<div class="BlockList-item is-dashed no-hover is-space u-tSpace-xl">
-  <div class="WidgetPlaceHolder">
-    <div class="WidgetPlaceHolder-media"></div>
-    <div class="WidgetPlaceHolder-txt">
-        <h2 class="WidgetPlaceHolder-title"></h2>
-        <p class="WidgetPlaceHolder-text"></p>
-    </div>
-  </div>
-  <button class="CDB-Button CDB-Button--primary js-add">
-    <span class="CDB-Button-Text CDB-Text is-semibold CDB-Size-small u-upperCase"><%- _t('editor.widgets.widgets-view.add_widget') %></span>
-  </button>
-</div>
+<h2 class="CDB-Text CDB-Size-huge is-light u-secondaryTextColor u-bSpace--xl"><%- _t('editor.widgets.widgets-form.placeholder-text') %></h2>
+<button class="CDB-Button CDB-Button--dashed CDB-Button--wide CDB-Button-Text CDB-Text is-semibold CDB-Size-small u-upperCase js-add">
+  + <%- _t('editor.widgets.add-widget.label') %>
+</button>
