+++ lib/assets/core/javascripts/cartodb3/editor/widgets/widgets-view.js
@@ -8,14 +8,19 @@ var widgetPlaceholderTemplate = require('./widgets-placeholder.tpl');
 var widgetsErrorTemplate = require('./widgets-content-error.tpl');
 var widgetsNotReadyTemplate = require('./widgets-content-not-ready.tpl');
 var checkAndBuildOpts = require('../../helpers/required-opts');
+var AddWidgetsView = require('../../components/modals/add-widgets/add-widgets-view');
+var TipsyTooltipView = require('../../components/tipsy-tooltip-view');
 
 require('jquery-ui');
 
 var REQUIRED_OPTS = [
   'userActions',
+  'analysisDefinitionNodesCollection',
   'layerDefinitionsCollection',
   'widgetDefinitionsCollection',
   'stackLayoutModel',
+  'userModel',
+  'configModel',
   'modals'
 ];
 
@@ -25,6 +30,10 @@ var REQUIRED_OPTS = [
 module.exports = CoreView.extend({
 
   module: 'editor:widgets:widgets-view',
+
+  events: {
+    'click .js-add': '_addWidget'
+  },
 
   initialize: function (opts) {
     checkAndBuildOpts(opts, REQUIRED_OPTS, this);
@@ -44,6 +53,12 @@ module.exports = CoreView.extend({
     this.clearSubViews();
     this.$el.empty();
 
+    this._initViews();
+
+    return this;
+  },
+
+  _initViews: function () {
     if (this._isLoading()) {
       this.$el.append(widgetsNotReadyTemplate());
     } else {
@@ -57,10 +72,18 @@ module.exports = CoreView.extend({
         } else {
           this.$el.append(widgetPlaceholderTemplate());
         }
+
+        var tooltip = new TipsyTooltipView({
+          el: this.$('.js-add'),
+          gravity: 'w',
+          title: function () {
+            return _t('editor.widgets.add-widget.tooltip');
+          },
+          offset: 8
+        });
+        this.addView(tooltip);
       }
     }
-
-    return this;
   },
 
   _initBinds: function () {
@@ -122,6 +145,24 @@ module.exports = CoreView.extend({
 
   _goToWidget: function (widgetModel) {
     this._stackLayoutModel.nextStep(widgetModel, 'widget-content');
+  },
+
+  _addWidget: function () {
+    if (this.$('.js-add').hasClass('is-disabled')) return;
+
+    var self = this;
+
+    this._modals.create(function (modalModel) {
+      return new AddWidgetsView({
+        modalModel: modalModel,
+        userModel: self._userModel,
+        userActions: self._userActions,
+        configModel: self._configModel,
+        analysisDefinitionNodesCollection: self._analysisDefinitionNodesCollection,
+        layerDefinitionsCollection: self._layerDefinitionsCollection,
+        widgetDefinitionsCollection: self._widgetDefinitionsCollection
+      });
+    });
   },
 
   _addWidgetItem: function (model) {
