+++ lib/assets/core/javascripts/cartodb3/editor/widgets/widgets-view.tpl
@@ -1,1 +1,4 @@
+<button class="CDB-Button CDB-Button--dashed CDB-Button--wide CDB-Button-Text CDB-Text is-semibold CDB-Size-small u-upperCase u-bSpace--m js-add">
+   + <%- _t('editor.widgets.add-widget.label') %>
+</button>
 <ul class="BlockList js-widgets"></ul>
